再提交一版,删除future结构,直接使用envelope作为等待队列数据,将所有的调用逻辑全放入消息总线模块中

+++ engine/actor/pid.go
@@ -15,7 +15,7 @@ func CreateServiceUid(serverId int32, serviceName, serviceId string) string {
 	return fmt.Sprintf("%d:%s@%s", serverId, serviceName, serviceId)
 }
 
-func NewPID(nodeUID, address string, serverId int32, serviceID, serviceName string, version int32) *PID {
+func NewPID(nodeUID, address string, serverId int32, serviceID, serviceName string, version int64) *PID {
 	if serviceID == "" {
 		serviceID = nodeUID
 	}
